Move test util function
